Refactor: Extract route filtering and page title to separate functions
